chore: update the exporting of services and web components in tooling and match the new pathing in sites
