+++ packages/tooling/fast-tooling/src/message-system-service/index.ts
@@ -1,1 +1,3 @@
 export * from "./ajv-validation.service";
+export * from "./monaco-adapter.service";
+export * from "./shortcuts.service";
