+++ packages/tooling/fast-tooling/src/message-system-service/monaco-adapter.service.spec.ts
@@ -9,11 +9,13 @@ import {
 import { mapDataDictionaryToMonacoEditorHTML } from "../data-utilities/monaco";
 import { MessageSystemSchemaDictionaryTypeAction } from "../message-system/message-system.utilities.props";
 import {
-    findDictionaryIdParents,
-    findUpdatedDictionaryId,
     MonacoAdapter,
     monacoAdapterId,
 } from "./monaco-adapter.service";
+import {
+    findDictionaryIdParents,
+    findUpdatedDictionaryId,
+} from "./monaco-adapter.service.utilities";
 import { MonacoAdapterAction } from "./monaco-adapter.service-action";
 
 chai.use(spies);
