+++ packages/tooling/fast-tooling/src/web-components/css-layout/index.ts
@@ -3,9 +3,9 @@ import { cssLayoutStyles as styles } from "./css-layout.styles";
 import { CSSLayout } from "./css-layout";
 
 /**
- * The FAST Tooling CSS layout Element.
+ * A web component used for updating CSS layout values.
  *
- * @public
+ * @alpha
  * @remarks
  * HTML Element: \<css-layout\>
  */
