+++ packages/tooling/fast-tooling/src/web-components/html-render-layer-inline-edit/html-render-layer-inline-edit.style.ts
@@ -1,6 +1,6 @@
 import { css } from "@microsoft/fast-element";
 
-export const HTMLRenderLayerInlineEditStyles = css`
+export const htmlRenderLayerInlineEditStyles = css`
     .edit-region {
         position: fixed;
         top: 0;
