+++ packages/tooling/fast-tooling/src/web-components/index.ts
@@ -1,2 +1,5 @@
 export * from "./color-picker";
 export * from "./css-layout";
+export * from "./html-render";
+export * from "./html-render-layer-inline-edit";
+export * from "./html-render-layer-navigation";
